+++ src/main/java/Duke.java
@@ -1,3 +1,9 @@
+import duke.Command;
+import duke.Parser;
+import duke.Storage;
+import duke.Task;
+import duke.TaskList;
+import duke.Ui;
 import java.util.Scanner;
 import java.nio.file.Path;
 import java.nio.file.Paths;
